feat: refactor source and committee handling, implement gRPC source for committee resolution
